Adds per-user installation support.  This fixes bug #17 (Install/uninstall release affects experimental instance of VS) and enables keeping one version of the tools installed per-user for dogfooding wihle still developing in the experimental hive.
When installing the user is 1st prompted with the EULA, then can choose install which does the default install, or they can choose Advanced which enables choosing Per-user or per machine install, it enables configuring the per-machine installation location (of our readme and short cut to VS), and it enables enabling or disabling features.

The choice of per-user or per-machine needs to occur before the CostInitialize /FileCost/CostFinalize so we need to have a bunch of "custom" UI which is the same as the normal Wix Advanced UI but it runs earlier in the sequences.  It also can't have back buttons because we'll run CostInitialize before we display the feature tree and navigating back at that point doesn't work.

The install is per-user in that it only touches files the user owns. But it does still require elevation.  Supporting both per-user and per-machine where per-user doesn't elevate isn't supported until Win7.  And even then policy is frequently set on machines in corporations which disallows non-elevated per-user installs (for example, my Srv2k8 R2 machine is configured this way by default).  Therefore we can't really take advantage of the MSIINSTALLPERUSER and ALLUSERS=2 feature.

Per-user installs are updated to include the appropriate per-user registry keys to enable the plugins inside of VS.

Also changes the way we disable installing IronPython/HPC features.  We now simply disable these by default rather than hiding them when the features are not installed.  This is done by setting our installlevel to 1 and raising their install level to 2 if the feature is not installed.  This allows the user to install these features in any order.

Also: fixes a bug in REPL w/ input and history.