Possible fix for bug #62 - Run-in-repl for custom interpreter deadlocks devenv.exe    I can't repro this bug but I can imagine where the deadlock is happening.  I think we're waiting to get the port number from the repl process (on the UI thread) and that we're never getting it for some reason.  So I've gone ahead and switched the direction of the socket connection so that VS provides the port number and waits for the REPL to connect back.  This also is a little bit simpler.
Also I've added some hardening for when we actually fail to start the REPL - this can easily be reproed by setting up a custom interpreter whos path doesn't start.  If you leave that REPL window open then close VS you actually will be unable to re-start VS w/o going in and deleting some registry keys.  This makes that succeed and reports the exception to the user that we couldn't start the process.  It also then handles the various things the ReplWindow can throw at us when we're in this disconnected state.

Finally I'm adding back provide auto load to the REPL window because it's failing to load in VS.  There's an underlying issue that needs to be investigated but this is a reasonable stop gap solution.  I've also stopped the REPL from eagerly asking for the scope list - instead it will nicely wait until the language tells it that the list is actually available which simplifies implementation in the REPL as it doesn't have to deal with not being connected yet.

I've also fixed an issue where the REPL doesn't display the proper list of scopes on startup.