90 - Crashes Visual Studio on certain kind of code in python file We need to not attempt to find a built-in module for a relative import where the module name is empty. Also add support for analyzing relative imports properly and an associated test case.
There's a performance problem when a large number of threads are created. We spend a bunch of time sending back a request for the thread's frame list, and then sending the frame list, while we're blocking the UI thread. To make it worse with lots of threads this starts to time out and then we start taking exceptions trying to send VS information about the frames. Now we just send the frame list whenever we are about to block the thread.

We also now report the thread ID as well as indicate whether a thread is the main thread or a worker thread.