511 Unable to load .pyproj in VS 2010 	Ignore the corrupt database file.
497 Problems with "New Project From Existing Code"
	Allow folders which start with .
	Also improve the error message so that we report so that it includes the bad filename

464 `mcs` is not recognised as a self-like variable in metaclasses
	We now process assignments to __metaclass__ (2x) and class keyword args which assign to metaclass (3x)


Plus delete some code that does nothing.

Fixes a perf and correctness problem which shows up w/ a very large project in particular w/ file->new project from existing code.  When we try and calculate the package name for a __init__.py we search parents to see if they also have __init_.pys.  But we’re accidently searching the entire hierarchy recursively and return the 1st one we find.  So we both search too much and that search can be wrong.

None of the fields of SliceINfo objects are currently being used in the abstract interpreter (although we do check for the type in some places) so just use a singleton for SliceInfo objects for the time being.