530 PTVS breaks Visual Studio Startup 	Handle the case of multiple intellisense controllers for the Python content type
120 Remove unused imports
	Adds remove unused imports feature and associated test cases.

This feature enables the user to remove unused imports from within a file.  This feature exposes two new commands, one for removing imports from the current scope and one for removing imports from all scopes.  These commands are exposed via the editor window context menu “Remove Imports” which has two sub-items “Current Scope” and “All Scopes”.  The “All Scopes” command is by default bound to the keyboard shortcut Ctrl-K Ctrl-O (knocking out those unused imports!).
This is similar to C#’s “Organize Usings” feature.  It differs in that there is no option to sort imports as their order can be significant within Python code while in C# their order does not matter.  We also add the feature to remove usings from either the current scope or all scopes as import statements can occur at different scopes and this is significant.
During the analysis this feature will only look at names which are imported and whether or not that name is used in any scope below where the import occurs.  There will be no accounting for control flow – for example using a name before an import statement will be treated as if the name was in fact used.
The analysis will ignore all “from __future__” imports, imports that are performed inside of a class definition, as well from from … import * statements.
1.1.1.1	Remove Imports -> Current Scope
The current scope command will determine the current scope based upon the location of the caret.  It will then identify all of the import and from import statements directly enclosed in the current scope.  It will then determine all of the identifiers used in the current scope and any children scopes.  Finally it will remove the imports or from import statements for any identifiers which were not used.
1.1.1.2	Remove Imports -> All scopes
In this mode the entire file will be processed for all imports.  First all imports used anywhere in the file will be determined.  Next all names which are referenced will be analyzed.  Finally any “import” and “from … import” statements whose names have not been used will be removed.
1.1.1.3	Removal
Removal of the imports will perform updates removing the imports while otherwise preserving the structure of the users’ code.  For import statements which include multiple modules being imported via a single import statement the import statement will be updated to remove any relevant imports.  If all imports are removed then the entire import statement will be removed while still preserving any leading or trailing comments.
Likewise for a “from … import” statement if multiple values are being imported then the unreferenced imports will be removed.  If this results in no imports occurring then the “from … import” statement will be removed in its entirety again while leaving any leading or trailing comments.
If the statements are removed in their entirety then it’s possible that an invalid program has been produced by having a missing empty statement in a function definition.  In this case a new empty “pass” statement will be inserted into the buffer where the import or from import statement has been removed.