49 Infer types from isinstance() calls 	Adds support for inferring types form isinstance() calls.  We support inferng types from both assertions, where the assertion applies until the end of the current scope, or within if blocks, where the check applies until the end of the if block.
Also a few cleanup/fixes:
Nonlocal keyword is being highlighted incorrectly
Find all refs on a variable used before an assignment statement shouldn’t refer to the variable in the outer scope, it should refer to the variable in the inner scope.
Fix how we’re getting the fully qualified module name when looking for relative imports.  It was correct for top-level imports, but not for nested imports.
SequenceInfo/IterableInfo had a duplicated return value field, always use IterableInfo’s instead