Fixes an issue when working w/ classes across files – when you remove a class, or move a class across files, we’re still keeping a reference to the old class around and offering it in goto definition and find all refs.  We now clear linked vars when re-parsing, re-enqueue dependents when clearing vars in a module, always track dependencies based upon the module currently being analyzed, and only give out located defs when they are still valid.
Fixes an off-by-one error where we were passing a 0-based line number from the VS APIs to a 1-based line number for our analysis APIs.

Fixes a test which is wrong and reporting that the analysis wasn’t up to date when it was .

Fix an issue where * and **args weren’t getting properly marked in the serialized database.

544 Cannot use REPL's stdout to provide output from child process's stdout
	Add fileno attribute to ReplOutput and create a pipe on demand which when written to will cause output to be displayed in stdout.
	Adds test case to ReplWindowTests.TestStdOutRedirected