591 PyKinect needs to work with Beta 2 sdk 	First pass on updates for the beta 2 SDK.  This is just the work required to get us up and running again, need to see if there’s more work to be done elsewhere.
	Update various paths for changes where things have moved in the SDK:
		DLLs now installed into System32
		PlanarImage now exposed via an interface w/ a couple of other structures to get info from the interface

587 StackOverflow when passing a format string to Console.Error
	We should do the formatting ourselves instead of recursing to death by doing the super call which calls back to us.

584 VS2010 SP1 crashes if folder renamed in Solution Explorer (not always)
	We have a file system watch which is listening for changes to the project directory.  It’s responsible for updating the icon for files when the files get deleted, renamed, etc…  on disk but not from within the project system.
	The problem is that we receive the events on a non-UI thread and aren’t currently dispatching back to the UI thread for handling them.  This means that we traverse the project hierarchy off the UI thread and call back into arbitrary VS components to notify them of changes – and all of this is expected to happen on the UI thread.

In 1.1 Alpha this is mitigated because we only do the traversal and never actually fire any re-draw events.  That’s because in 1.1 we’ve made sure that folder paths are normalized and always end in \’s, and therefore we never find the folders, and never delivery any change events.

The fix is to set an ISynchronizeInvoke on the file system watcher so that we post messages back onto the UI thread.  We also no longer deliver the file change events at all when we’re performing the rename from within Visual Studio.  And finally we now remember to get rid of the file change notification after the project is unloaded.


Sync pyvot to latest version, including tests and docs. setup.py as used by VS-based installation is now the same one as used normally; a special flag has been added to prevent dependency downloads when invoked from VS.