Adds support for referencing .NET projects and assemblies from a Python project.
We add a new ProjectAssemblyReference class for telling the interpreters about the reference, and a new PythonProjectReferenceNode which handles delivering the notifications to the interpreter when assemblies are removed or added.  Users can now add references directly to an existing .NET assembly or they can add references to a project.  Finally the IronPythonInterpreter will also use the search paths to find assemblies.

Removes the unused Project reference for the AddReferenceAsync API – instead we always give the interpreter the real assembly when it’s available, and we handle tracking the file changes so the interpreter doesn’t need to deal w/ VS at all.

Includes a bunch of changes regarding thread safety and documenting what stuff is thread safe.  Some dictionaries get switched to concurrent dictionaries so we can access them from any thread.  Also starts using more of the new .NET Task stuff including adding our own synchronization context for our analysis thread so that we can easily capture our context and push it back onto the analysis thread.

This also allows  us to now queue analyzing referenced directory off the UI thread – previously there was a big potential UI blocker that could pause for some time depending on how many files are in the entire directory.

Kill CommonReferenceNode – this was new in 1.1, and came from the old IronStudio effort, but it brings nothing to the party, so we’ll inherit from its base class instead.

Adds test cases for both project and assembly references.

Misc fixes:
	Fix spelling of CrossModuleAnalysisLimit