610 Create project in existing folder 	Changes our templates so that we don’t require a new directory – templates still get created in the same way, the directory can just already exist.
Fixes internally reported issue w/ emacs mode – set PrimaryDocument on the REPL role as we are the primary view of the current document (not really sure why Emancs mode doesn’t want to be enabled on secondary views).
Also changes our templates so the first .py file is named the same as the project/solution name.