Writing to the stdout file handle is not reflected in the REPL window once it's started up.
Discovered this while looking at http://pytools.codeplex.com/workitem/608

When we spawn a subprocess, and it writes to the stdout file handle, this doesn't show up in the REPL window when it should.  Switch to using the OutputDataReceived/ErrorDataReceived events which deliver the output line by line (versus reading w/ a buffer of size 1 so that we get all of the output).  Also saves us creating a couple of threads.

Also fixes the startup items on our projects to use the project filename.