645 Attaching can leave non-running thread blocked
When we do the attach one thread is blocked in native code.  We attach, resume execution, and that
thread should eventually wake up.

The bug was two issues, when doing a resume all:
	1) we don't clear the stepping if it's STEPPING_ATTACH_BREAK
	2) We don't clear the stepping if we haven't yet blocked the thread

Because the thread is blocked in native code, and we don't clear the stepping, when the user
hits resume the thread will eventually return back to Python code, and then we'll block it
because we haven't cleared the stepping bit.

	The fix is to clear our stepping state even if we’re not yet blocked and to clear our stepping state for break or an attach break.