Fix an issue w/ extract method where we don’t walk the statements in outer suites that follow after we’ve extracted from an inner suite.  The code:
class C:
    def foo(self):
        if False:
            bar = player = Player()

        player.update()


When extracting “bar = player = Player()” we’re not seeing the call to “player.update()” because we don’t walk that code, we only walk the non-existent statements immediately after the assignment.

Now we keep track of the statements as well as the funcdef/class def scopes, and include any of the following statements to be walked.

Add test case to ExtractMethodTests.AssignInIfStatementReadAfter