Another extract method dogfooding bug.
When walking the LHS of an assignment we walk it for definitions.  But it’s only really a definition if it’s going to be a NameExpression – otherwise it’s an assignment into a member or index of some variable which already exists.  Ultimately we need to walk that variable for reading so that we’ll properly include it in the list of variables which need to be flowed in.