More Variable analysis and perf improvements
More tweaks to extensible variable analysis and how Django plugs in to this. Tweaks ExternalAnalysisValue so it supports Call and GetMember participation.

After this change the code:

 latest_poll_list = Poll.objects.all().order_by('-pub_date')[:5]
 t = loader.get_template('polls/index.html')
 c = Context({
 'latest_poll_list_foo': latest_poll_list,
 })
 return HttpResponse(t.render(c))

Is successfully analyzed and we know that polls/index.html should have the latest_poll_list_foo variable available to it.

Also tweaks our Cartesian product function analysis so that locals (including isinstance locals) are now fully separated between analysis in addition to parameters.

Perf:
 A bunch of tweaks to reduce overhead in various places where we're copying sets.

TypeUnion now internally switches from null->SetOfOne->SetOfTwo->HashSet. When returning the set we don't need to copy the empty set/set of one/set of two because they're immutable. Also makes TypeUnion implement ICollection<T> so that when we create a set from it we get more efficient copying based upon the # of elements in the set.

 VariableDef more efficiently enumerates over the dependencies and avoids intermediate copies when building up the final resulting set. It also accesses EmptySet<T> less (which turns into a moderately expensive lookup in the CLR)

 SingleDict.Values avoids creating an intermediate enumerator, and adds SingleDict.DictValues for when we know we have a dictionary so we get the strongly typed ValueCollection.

 Altogether this yields around a 20% perf gain when running against a fairly basic Python 2.6 install (from 70 to 56 seconds on my machine).

Also:
 Fixes a thread safety issue w/ function specialization – we were racing against the parsing of the Django code and wouldn't always get our specialized functions working.
 Removes some dead code from DDG
 Fixes a span in TemplateClassifier
 Doesn't include empty strings in QuickInfo (for our external analysis values)
 Bug 686 - IronPython not picked up if installed after PTVS
 Removes the check to see if Ipy is installed so IronPython support is always available by default. Users can disable it.