735 PTVS, PyPy and Django app 	Scraping of the std lib on PyPy is messed up when we encounter plain old Python function objects.  This causes problems both with simple functions as well as methods stored in a type.
	For methods we can see that things like “list.__dict__['append']” are actually just function objects in PyPy.  On other implementations these are “method_descriptor” objects which are distinct from both built-in functions and user defined method objects.

The fix is to update our scraper so that it recognizes Python function objects and so that it understands that a function object stored in a type should be treated like a method.

There’s also another issue in that we fail to find PyPy’s lib dir, so I’ve also added some logic to find the lib dir by looking at where site.py is located from the running interpreter.

Reving the analysis version so that users will automatically have their DB regenerated.