Adds code re-formatting support to PTVS with a large set of options.
For the most part this is updating the parser's round tripping support to handle using a CodeFormattingOptions object to override the pre-existing white space.

A few tweaks were necessary to how we handle string plus and function definitions for their white space. For function definitions this propagates out as changes to the outlining support and the extract method functionality. It also fixes a bug in extract method.

The options UI in Tools->Options is implemented using reflection over the options. Adding a new option and getting it into the UI is easy – add it to CodeFormattingOptions, and add the 4 attributes. If you want to a new category, add the new value to the CodeFormattingCategory enum, and then add a new page or add a “new OptionCategory(YourCategory, OptionCategory.GetOptions(CodeFormattingCategory.YourCategory))” to an existing page. The options page is modeled on C#'s options pages with some accessibility code taken directly from the C# code base.

Includes lots of test cases which run outside of VS for the individual formatting. Also includes handful of tests to make sure the feature works end-to-end in VS include hitting the tools->options page, making sure that we correctly apply the minimal edits to the text buffer, and that the actual commands are wired up.

Also:
Seperates out the mutate std lib tests into per-version variants, and adds test inconclusive if interpreter isn't installed.
Sets the ConditionalBP metric which is missing in the debugger.