These changes get the DebuggerUI tests running successfully against 2.* and 3.* interpreters.
Most changes are to the tests. There are product changes to fix the bug that warnings stick around in the TaskList window after project close. This was causing interactions between DebuggerUI tests run in batches.

With these changes all DebuggerUI tests pass when run individually and all but one pass when all are run in batch. TestAttachThreadsBreakOneAndSetExitFlag fails when run in batch. I have narrowed down the failure to an interaction with the TestException* tests, but do not yet have a fix.