Fixes the Django non-UI tests.
We were using a .yaml file to read in the data, but that requires PyYAML to be installed. Switch to JSON which works by default.

The YAML file also included some information that we were unable to import due to conflicts after doing syncdb. But this json file successfully imports for the time being.