Fix a couple of more code formatting issues.
We don't handle line continuation characters correctly because we reflow them, so:

 if a and \
 b:
 pass

Gets an extra line in it.

We also don't handle empty selections correctly and we end up making a negative span instead of doing nothing.
Also fixing extract method for the same thing.

We also don't handle changing prefixes, such as:

 #
 # foo
 # bar

Where the comment prefix for the 1st line is   # and additional lines it is   #  , so now we do a stricter comparison for prefix lines.

Also adds a hint path for Microsoft.Windows.Design.Host, which was giving me issues when building for Dev10 last night.