1174 Interpreter list should resize first column when new interpreters detected Adds auto resize of the first column after interpreters are refreshed.
1198 Interpreters without completion DBs should display a different message in interpreter list
Adds new label for when an interpreter does not support refreshing. This may be because the factory does not implement IInterpreterWithCompletionDB or because the interpreter path is invalid.
Changes TextBlock elements to Labels to make them visible to automation tools.

1199 fake interp path kills VS
As well as the fix for 1198 (which hides the Refresh DB button), also adds exception handling for database generation. This will improve further when interpreters are (partially) refactored into their own assembly.

Also improves the combo box selections in the options controls after deleting an interpreter and adds extra DEBUG text to the entries.
Removes Generate DB button from options dialog.
Adds tests for interpreter window.