#980 Rename of folder fails when item is missing
Make renaming a non-existing file project item a no-op in terms of disk operations (rather than trying and producing an error that the original doesn't exist). This abandons the previous logic that only attempted to remain silent when the destination file of the rename was physically present. If we do not complain about the rename of a folder with a non-existing file (which involves renaming that file), there is no reason to complain about renaming the file directly. Furthermore, we allow renaming non-existent folders regardless of whether the source or the target exists. This change makes the behavior consistent for all these cases.