1741 General text formatting options are not saved Saves the Paste removes REPL prompts option. Disables Automatically format on paste since it is not implemented. Also removes unnecessary groupbox and adds layout control.
1743 Python installation while VS was running was not detected
Fixes two race conditions and prevents a potential crash.
The first race could occur when Python is being installed for the first time. We would detect that 'Software\Python' has been created, but if PythonCore had not been created by then we would never change to watching it. We now change to watching Python in this case and then to PythonCore when it becomes available (this is cheaper than recursively watching the entire Software hive).
The second race could occur between the DiscoverInterpreterFactories() call and starting the watcher on PythonCore. The order of these calls is now reversed to avoid this.
Finally, a crash would occur if the key being watched was removed. We now detect this case and switch back to watching the parent keys. We also handle exceptions thrown when registering keys in RegistryWatcher and use Debug.Fail() instead of crashing.