#2030 Invalid search path crashes VS Filters invalid search paths out before attempting to analyze them. This ensures that they appear in Solution Explorer (as missing) but are not analyzed if they do not exist. Also prevents search paths from being added to the list of files for Find in Files.
Enables adding ZIP archives as search paths on VS 2010 (with .NET 4.5) and fixes removal of search paths.

Changes to use more reliable mechanism for bolding items on VS 2012 and later. Rather than waiting for the idle-loop, we listen to the IVsHierarchyItemManager.OnItemAdded event. The old code remains unmodified for VS 2010.