#2093 Debug unit test on VS2013 gives error dialog, then VS hangs
This deals with the then VS hangs (on protocol version mismatch). There are two problems here. The first one is that when we have a mismatch, we immediately drop the socket on VS side, leaving ptvsd hanging. This changes it to send the complete answer back before dropping the socket, so that ptvsd can gracefully close it on its own end and start waiting for the next connection.

The second problem is that EnumProcesses was not correctly returning an error HRESULT on error, but rather success with an empty collection. Furthermore, even if error is returned, the debugger DTE objects do not correctly propagate it. As a result of that, the unit test adapter code does not receive any errors when attempting to attach to process, and keeps retrying. This adds correct error handling to EnumProcesses, and also adds a validation to AddPort, which debugger DTE objects do correctly handle and propagate.