#2132 Attach to Process is broken when old version of ptvsd is in site-packages Change the order in which modules are probed such that the attach case is tried first, and ptvsd is last.
#1953 Debug attach - startup script is not listed in modules window
#2133 Modules that are loaded before the main script starts running do not show up in Modules window
Change module reporting logic to always report all modules after debugger connection is established. This is a behavioral change, but the old behavior (do not report modules loaded before startup script, but do report modules loaded after) was non-obvious, and in case of attaching, it also precluded the main user script (and anything it imported) from showing up, which was reported as a bug.

Added a change to SafeRepr such that if merely reading the object type's __repr__ throws, we let that propagate all the way to the top-level exception handler which will print out an exception, rather than saying object repr not available - on the grounds that this is an abnormal situation that user should have more details about.