#2545 Step Into on print ends up in socket.sendall Fix check for internal debugger calls to standard library.
#2514 Mixed mode debugging does not support Managed 2.0/3.0/3.5 code type
Mark the corresponding code type as incompatible with Python

#1958 Remote debug 64-bit python on linux outputs errors to debug interactive
Handling of 64-bit thread IDs in the REPL was generally broken, because it did not do any ID mapping at all. Expose mapping on AD7Engine via an interface, and use it in the REPL.

Fix breakpoint on the first line to execute not working when attaching.

Fix EvalRawTest on IronPython.

Add DebugReplEvaluatorTests for new Python 3.x versions.