#2811 PyKinect documentation page needs to refer to samples vsix - Port to markdown, update for PTVS 2.1. Current docs were still for 2.0.
Fixes parsing of video links to allow for video ids with a dash '-' in them. This was the case for the flask+mongodb video, which wasn't getting processed.

Fixes the list_outputs_only phase, which is used by the PTVS docs pyproj for converting the markdown to html at build time.

Updates all our videos to render at 0.5 scale, so that the thumbnail isn't so large.