Don't add empty search paths to the environment.
Previously, the code would construct a key,value pair both consisting
of empty strings, which in turn would cause an empty entry to be
added to the environment block used to launch the process.

This would result in VsShellUtilities.LaunchProcess() failing
internally when launching the interpreter (and it would print
"Exception : Could not launch <python-interpreter>" to the debug
console.

With this patch, we just do nothing if the search paths are empty.