Fix #132: Debugger options are ignored when attaching
Propagate all global debug options to the debug engine whenever it is created in a uniform way (in all attach and launch scenarios), and from there to the debugger script.

Use bitflags instead of textual switches to propagate flags to the debugger script for easier parsing and future extensibility.

Bonus feature: ability to specify additional debug options (e.g. enable Django debugging) in the URL when attaching remotely via ptvsd.

Add tests for output redirection for launch and attach scenarios to test flag propagation.

Fix #259: Fully qualified function names in Call Stack are computed incorrectly if there is a source mismatch

If the computed fully qualified name does not correspond to the unqualified function name reported by the runtime, fall back to the latter.