Propagate debug options as string rather than bitflags when passing it to the Python script.
Fix 64-bit attach and wait-on-exit that were broken by previous debug options changes.

DebuggerTests is now an abstract class, and DebuggerTests26 is added for 2.6 tests. Test classes for 2.7, 3.3 and 3.4 are separated into 32-bit and 64-bit versions so that both are tested when doing a full test run (all other Python versions will still be tested using whatever is installed, to avoid running too many tests).