Fix #301: AttachWithOutputRedirection test asserting from incorrect thread
Rewrote the test to only capture output while running, and assert afterwards.

Fixed bug exposed by the updated test with Python 3.4 and 3.5 where directly attaching to the process that hasn't imported threading module yet would cause an assert in that module to fail on process shutdown because it incorrectly identified a debugger helper thread as a main thread.

Changed the use of local imports in conjunction with 'global' to not rely on implementation details of that interaction.