Fix #1248    Really this is just a partial fix, but it fixes the reported issue of    the smart tag not coming back.  This brings in changes made in NTVS to    VariableDef, in particular the fix for this is that in AddTypes we no    longer use value.DeclaringModule to figure out where to store the    value.  Instead we always assign based upon the project doing the    assigning.  Unfortunately this means that we can still propagate types    around across modules.
Due to bringing in the VariableDef changes we also start to pick up some
infrastructure for cross-module visibility features.  This is not yet
enabled because we need to come up with a general policy around when to
make things visible.  But most of the callers into VariableDef are updated
to support flowing in the appropriate state for doing the filtering.

Also some cleanup of iterable/iterator classes to handle cases where
they're not tied to VariableDef's and therefore won't support filtering.