Fix #727 Adds an aggregatable project entry concept which allows dependency info to be tracked based upon the versioning of multipl project entries.
To accomplish this IProjectEntry needs to have AnalysisVersion split off
to a specialized IVersioned interface which is what
DependentData (and CallChainSet) use to do their tracking of data.  When
we go to get the tracking project entry we now go to
AnalysisUnit.DependencyProject instead of ProjectEntry.
FunctionAnalysisUnit's are split out into the non-called ones and the
called ones which have call chains, and only the called ones override the
DependencyProject with one which handles multiple
project entries with dependencies.

Previously we would use DependentData's project entry to provide file
information for our encoded locations.  That information
is now lost, so our existing LocationInfo, IReferencable, EncodedLocation,
etc...  all need to make sure that they can provide
the file path themselves.  We now make sure that our location resolver
holds enough context to resolve the path by moving that
to AnalysisUnit instead of the PythonAst's.