Added file watcher for the site-package directory.
Msbuild item groups are only evaluated at project load.  Because of this, even if new files were added to the site-package directory, msbuild would not re-evaluate the item group -- "$(MSBuildProjectDirectory)\Interpreter\lib\site-packages\**\*".  Thus, the new files would not be included in the appx.  This file watcher is to force msbuild to re-evaluate its item group whenever the site-package directory is modified.