Fixes #1906 New custom environment does not land user in configure Fixes #1943 Creating Python environment forcibly prepends it with "VisualStudio" Fixes #1944 Changing environment library path does not toggle Apply button
Reworks how environments are created so that we don't eagerly save any details. Allows much easier cancellation.
Removes the UI for Library Path
Fixes the registry key where we save the display name.
Auto-detect will now fill in the description
Adds a button to the Overview tab to jump to the configure panel.