pytest not found in azure-cli Fix #5626
Issues: Pytest execute will search up the parent folders for a config .ini file which will change the rootdir from the project to the subfolder.  This will cause a mismatch between discovered test filepaths and results.

now force both disocvery  and execution to set the rootdir so that .ini location wont override it.

also note it appears the user can't set the rootdir from inside the .ini file
ie
 addopt = --rootdir=XXX
 or
 rootdir = XXX

both appear to have no effect so it should be safe to force it to be the project/workspace location.

Users can still set "testpaths = testing" to optimize the search but the paths will be relative to the project/workspace dir from now on.

https://docs.pytest.org/en/2.8.7/customize.html