Fix issue with executing the activate_this script
The activate_this.py file that is being executed by get_virtualenv_handler() might set the __file__ attribute to a unicode value. If I'm not mistaken then the value of __file__ should be of type 'str' in Python 2.7.

I discovered this when I ran into an issue where a core Python2.7 script (FixTk.py) raised an AssertionError because sys.prefix was of type unicode. It turned out that the activate_this.py script was replacing sys.prefix with a value based on its __file__ attribute.

I added 2 LOC to call the already defined to_str() function on the value of 'activate_this'.