fix for narrator in packages windows.
Now forcing a selection of label "IsDefaultMessage" after button press, and triggering narrator to update in the code behind so that it reads "This is the default environment for new projects".

fix internal bug 1351252

Actual Result:​
Narrator/NVDA is not announcing the status (This is the default environment for new projects) which is available on screen on selecting 'Make this the default environment for new projects' button.

Expected Result:​
Narrator/NVDA should announce the status as "This is the default environment for new projects" after selecting 'Make this the default environment for new projects' button.