Fix text/enabled state of new project/file commands (#8295)
* Fix "From Cookiecutter..." text in Dev18

There is a new drop down on the toolbar where this command shows up, and its text does not fit in with the rest of the commands.

This changes this command to mimic the rest of the commands attached to that drop down.

This also changes its non-descriptive canonical name from "File.CreateFrom" to "File.NewFromCookiecutter"

* Disable commands when building/debugging

Similar to the equivalent commands in Shell, disable these commands while building and debugging.

* add missing break

---------

Co-authored-by: Stella Huang <stellahuang@microsoft.com>